refactoring out exceptions so they can be imported elsewhere
